+++ tests/__init__.py
@@ -21,7 +21,7 @@ TEST_SAMPLE_DATA = os.path.join(_HERE, "data")
 def uri_to_path(uri):
     parsed = generous_parse_uri(uri)
     if IS_WINDOWS:
-        return unix_to_windows_path(parsed.path, parsed.netloc)
+        return unix_to_windows_path(parsed.path)
     return parsed.path